Tags Commit

+++ app/Http/Controllers/ProductController.php
@@ -3,13 +3,15 @@
 namespace CodeCommerce\Http\Controllers;
 
 use CodeCommerce\Category;
-use CodeCommerce\Http\Requests;
 use CodeCommerce\Http\Requests\ProductsRequest;
+use CodeCommerce\Tag;
 use CodeCommerce\Product;
+use CodeCommerce\Http\Requests;;
 use CodeCommerce\ProductImage;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Facades\File;
+use League\Flysystem\AwsS3v3\AwsS3Adapter;
 
 
 class ProductController extends Controller
@@ -22,18 +24,17 @@ class ProductController extends Controller
 
     private $productModel;
 
-    public function __construct(Product $productModel)
+    public function __construct(Product $productModel,Tag $tagModel )
     {
         $this->middleware('guest');
-
         $this->productModel = $productModel;
+        $this->tagModel = $tagModel;
     }
 
     public function index()
     {
 
         $products = $this->productModel->paginate(10);
-
         return view('products.index', compact('products'));
     }
 
@@ -46,71 +47,69 @@ class ProductController extends Controller
     {
 
         $categories = $category->lists('name', 'id');
-
-
-
         return view('products.create', compact('categories'));
     }
 
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @return Response
-     */
     public function store(ProductsRequest $request)
     {
+        $inputTags = explode(',', $request->get('tags'));
+
         $input = $request->all();
 
         $input['featured'] = isset($input['featured']) ? 1 : 0;
         $input['recommend'] = isset($input['recommend']) ? 1 : 0;
 
-        $product = $this->productModel->fill($input);
+        $inputTags = explode(',', $input['tags']);
+        $newTagsRelatedToProduct = $this->extractTagsFromInput($inputTags);
 
+        $product = $this->productModel->fill($input);
         $product->save();
+
+        $product->tags()->attach($newTagsRelatedToProduct);
+
 
         return redirect()->route('products');
 
     }
 
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  int  $id
-     * @return Response
-     */
     public function edit($id, Category $category)
     {
-        $categories = $category->lists('name', 'id');
-
         $product = $this->productModel->find($id);
+        $categories = $category->lists('name', 'id');
 
         return view('products.edit', compact('product', 'categories'));
 
     }
 
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  int  $id
-     * @return Response
-     */
     public function update(Request $request  ,$id)
     {
 
-        $this->productModel->find($id)->update($request->all());
+        $input = $request->all();
+
+        $inputTags = explode(',', $input['tags']);
+
+        $newTagsRelatedToProduct = $this->extractTagsFromInput($inputTags);
+
+        $product = $this->productModel->find($id);
+        $product->update($input);
+
+        $product->tags()->sync($newTagsRelatedToProduct);
 
         return redirect()->route('products');
 
     }
 
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  int  $id
-     * @return Response
-     */
     public function destroy($id)
     {
+        $imagesToDelete = $this->productModel->find($id)->images;
+
+        foreach($imagesToDelete as $image) {
+
+            Storage::disk('public_local')->delete($image->id . '.' . $image->extension);
+
+            $image->delete();
+        }
+
         $this->productModel->find($id)->delete();
 
         return redirect()->route('products');
@@ -139,7 +138,7 @@ class ProductController extends Controller
 
         Storage::disk('public_local')->put($image->id.'.'.$extension, File::get($file));
 
-        return redirect()->route('products.images',['id'=>$id]);
+        return redirect()->route('products.images', ['id'=>$id]);
 
     }
 
@@ -147,13 +146,36 @@ class ProductController extends Controller
     {
         $image = $productImage->find($id);
 
-        Storage::disk('public_local')->delete($image->id.'.'.$image->extension);
-        
-        $products = $image->product;
+        if(file_exists('/uploads/'.$image->id.'.'.$image->extension)) {
+            Storage::disk('public_local')->delete($image->id.'.'.$image->extension);
+        }
+
+        $product = $image->product;
         $image->delete();
 
-        return redirect()->route('products.images', ['id'=>$products->id]);
+        return redirect()->route('products.images', ['id'=>$product->id]);
         
         }
+
+    private function extractTagsFromInput($inputTags)
+    {
+        $newTagsRelatedToProduct = [];
+
+        foreach ($inputTags as $tag) {
+            $tag = strtolower(trim($tag, ' .;-!?#$&@*()|+=_{}[]^~'));
+
+            $newTag = new Tag(['name' => $tag]);
+
+            $tagAlreadyExists = $this->tagModel->where('name', '=', $tag)->first();
+
+            if (!isset($tagAlreadyExists)) {
+                $newTag->save();
+                array_push($newTagsRelatedToProduct, $newTag->id);
+            } else {
+                array_push($newTagsRelatedToProduct, $tagAlreadyExists->id);
+            }
+        }
+        return $newTagsRelatedToProduct;
+    }
 
 }
